Refactor pages with utility helpers
