+++ application-system/src/main/java/com/group_9/project/PlansPage.java
@@ -4,13 +4,11 @@ import com.group_9.project.session.UserApplicationData;
 import com.group_9.project.utils.*;
 
 import javax.swing.*;
-import javax.swing.plaf.basic.BasicScrollBarUI;
 import java.awt.*;
 
 public class PlansPage extends JFrame {
     public PlansPage() {
-        ImageIcon icon = new ImageIcon(getClass().getClassLoader().getResource("images/app_icon.png"));
-        setIconImage(icon.getImage());
+        BaseFrameSetup.applyAppIcon(this);
         BaseFrameSetup.setupFrame(this);
 
         JPanel mainContainer = new JPanel();
@@ -61,7 +59,7 @@ public class PlansPage extends JFrame {
         scrollPane.setBorder(null);
 
         JScrollBar verticalScrollBar = scrollPane.getVerticalScrollBar();
-        verticalScrollBar.setUI(new CustomScrollBarUI());
+        verticalScrollBar.setUI(new ScrollUtil.PurpleScrollBarUI());
         verticalScrollBar.setOpaque(false);
         verticalScrollBar.setPreferredSize(new Dimension(10, 0));
 
@@ -214,45 +212,6 @@ public class PlansPage extends JFrame {
         return card;
     }
 
-    private static class CustomScrollBarUI extends BasicScrollBarUI {
-        private static final Color THUMB_COLOR = new Color(42, 2, 67);
-
-        @Override
-        protected void configureScrollBarColors() {
-            this.thumbColor = THUMB_COLOR;
-            this.trackColor = new Color(0, 0, 0, 0);
-        }
-
-        @Override
-        protected JButton createDecreaseButton(int orientation) {
-            return createZeroButton();
-        }
-
-        @Override
-        protected JButton createIncreaseButton(int orientation) {
-            return createZeroButton();
-        }
-
-        private JButton createZeroButton() {
-            JButton button = new JButton();
-            button.setPreferredSize(new Dimension(0, 0));
-            return button;
-        }
-
-        @Override
-        protected void paintThumb(Graphics g, JComponent c, Rectangle thumbBounds) {
-            if (!scrollbar.isEnabled() || thumbBounds.isEmpty()) return;
-            Graphics2D g2 = (Graphics2D) g.create();
-            g2.setColor(THUMB_COLOR);
-            g2.fillRoundRect(thumbBounds.x, thumbBounds.y, thumbBounds.width, thumbBounds.height, 10, 10);
-            g2.dispose();
-        }
-
-        @Override
-        protected void paintTrack(Graphics g, JComponent c, Rectangle trackBounds) {
-            // No track painting
-        }
-    }
 
     public static void main(String[] args) {
         SwingUtilities.invokeLater(PlansPage::new);
